refactor(connectors): give the local password DB its own file, drop the server aliases
Split passwordDB into connectors/password.go, leaving resolve.go with just the
resolver and the config-map contract. Remove the server-package type/const
aliases to connectors.ConnectorConfig / connectors.LocalConnector; cmd now
references those symbols in the connectors package directly. server keeps the
built-in ConnectorsConfig map (typed with connectors.ConnectorConfig) that it
hands to connectors.Resolver.
