+++ server/server_grant_password_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"golang.org/x/crypto/bcrypt"
 
+	"github.com/dexidp/dex/server/connectors"
 	"github.com/dexidp/dex/storage"
 )
 
@@ -117,7 +118,7 @@ func TestHandlePassword_LocalPasswordDBClaims(t *testing.T) {
 	// Enable local connector.
 	localConn := storage.Connector{
 		ID:              "local",
-		Type:            LocalConnector,
+		Type:            connectors.LocalConnector,
 		Name:            "Email",
 		ResourceVersion: "1",
 	}
